[Standalone] Fixed lingering TempDirToucher. Closes GH-1678.
The TempDirToucher will now periodically reduce its privileges, which allows it to be killed when passenger is quit if the watchdog has also reduced its privileges.

The watchdog cd’s to the temp dir, so that when it drops privileges it can still read the TempDirToucher’s pid file in the temp dir (parts of the path to the secure TMPDIR are missing o+x in some cases).